pyproject.toml
[tool.poetry]
name = "todo"
version = "0.1.1"
description = ""
authors = ["Уманец Елизавета <e.umanets@zebrains.team>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.63.0"
tortoise-orm = "^0.16.21"
pydantic = "^1.8.1"
uvicorn = "^0.13.4"
asyncpg = "^0.22.0"
aerich = "^0.5.1"
celery = "^5.0.5"
redis = "^3.5.3"
fastapi-admin = "^1.0.1"
aioredis = "^1.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
